+++ src/main/java/com/mzcteam01/mzcproject01be/domains/lecture/controller/OfflineLectureController.java
@@ -6,7 +6,7 @@ import com.mzcteam01.mzcproject01be.domains.lecture.dto.response.LectureOfflineD
 import com.mzcteam01.mzcproject01be.domains.lecture.dto.response.LectureOfflineListResponse;
 import com.mzcteam01.mzcproject01be.common.exception.LectureErrorCode;
 import com.mzcteam01.mzcproject01be.domains.lecture.enums.SearchType;
-import com.mzcteam01.mzcproject01be.domains.lecture.service.facade.interfaces.LectureService;
+import com.mzcteam01.mzcproject01be.domains.lecture.service.interfaces.LectureService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
