+++ src/main/java/com/mzcteam01/mzcproject01be/domains/lecture/service/interfaces/LectureFacadeService.java
@@ -1,4 +1,4 @@
-package com.mzcteam01.mzcproject01be.domains.lecture.service.facade.interfaces;
+package com.mzcteam01.mzcproject01be.domains.lecture.service.interfaces;
 
 import com.mzcteam01.mzcproject01be.domains.lecture.dto.response.GetLectureResponse;
 import com.mzcteam01.mzcproject01be.domains.lecture.entity.Lecture;
