+++ src/main/java/com/mzcteam01/mzcproject01be/domains/lecture/service/interfaces/OfflineLectureService.java
@@ -1,4 +1,4 @@
-package com.mzcteam01.mzcproject01be.domains.lecture.service.facade.interfaces;
+package com.mzcteam01.mzcproject01be.domains.lecture.service.interfaces;
 
 import com.mzcteam01.mzcproject01be.domains.lecture.dto.response.LectureOfflineDetailResponse;
 import com.mzcteam01.mzcproject01be.domains.lecture.dto.response.LectureOfflineListResponse;
